# http-api/rest_api/swagger.json
{"swagger":"2.0","info":{"title":"OX REST API","description":"Documentation of the Open-Xchange REST API.\n\n___\n","contact":{"name":"Open-Xchange GmbH","email":"info@open-xchange.com","url":"https://www.open-xchange.com/"},"version":"7.8.4"},"host":"example.com","schemes":["http","https"],"basePath":"/","tags":[{"name":"Userfeedback","description":"The user feedback module"},{"name":"Advertisement","description":"The advertisement module"}],"securityDefinitions":{"basicAuth":{"type":"basic","description":"HTTP Basic Authentication."}},"paths":{"/userfeedback/v1/{context-group}/{type}":{"delete":{"operationId":"delete","tags":["Userfeedback"],"summary":"Deletes user feedback","produces":["application/json"],"parameters":[{"in":"path","name":"context-group","description":"The context group identifying the global DB where the feedback is stored.","required":true,"type":"string"},{"in":"path","name":"type","description":"The feedback type to delete.","required":true,"type":"string"},{"name":"start","in":"query","description":"Start time in milliseconds since 1970-01-01 00:00:00 UTC. Only feedback given after this time is deleted. If not set, all feedback up to -e is deleted.","required":false,"type":"integer","format":"int64"},{"name":"end","in":"query","description":"End time in milliseconds since 1970-01-01 00:00:00 UTC. Only feedback given before this time is deleted. If not set, all feedback since -s is deleted.","required":false,"type":"integer","format":"int64"}],"responses":{"200":{"description":"Info message when successfull.","schema":{"type":"string"}},"400":{"description":"Bad request, response contains error message.","schema":{"type":"string"}},"401":{"description":"Not authorized"},"404":{"description":"Not found, if path is wrong (unknown context group or feedback type). Response contains error message.","schema":{"type":"string"}},"500":{"description":"Internal server error that might have multiple reasons, for instance no configured global database. Response contains error message.","schema":{"type":"string"}}},"x-ox-examples":{"Example":{"description":"Delete user feedback from 2016-01-01 to 2016-12-31","example":"DELETE http://localhost:8009/userfeedback/v1/default/star-rating-v1?start=1451606400&end=1483228799"}}}},"/userfeedback/v1/export/{context-group}/{type}":{"get":{"security":[{"basicAuth":[]}],"operationId":"exportCSV","tags":["Userfeedback"],"summary":"Exports user feedback as CSV","produces":["application/octet-stream"],"parameters":[{"in":"path","name":"context-group","description":"The context group identifying the global DB where the feedback is stored.","required":true,"type":"string","default":"default"},{"in":"path","name":"type","description":"The feedback type to send.","required":true,"type":"string","default":"star-rating-v1"},{"name":"start","in":"query","description":"Start time in milliseconds since 1970-01-01 00:00:00 UTC. Only feedback given after this time is sent. If not set, all feedback up to -e is sent.","required":false,"type":"integer","format":"int64"},{"name":"end","in":"query","description":"End time in milliseconds since 1970-01-01 00:00:00 UTC. Only feedback given before this time is sent. If not set, all feedback since -s is sent.","required":false,"type":"integer","format":"int64"},{"name":"delimiter","in":"query","description":"The column delimiter used. Default: ';'","required":false,"type":"string"}],"responses":{"200":{"description":"Export data into a CSV file","schema":{"type":"file"},"headers":{"Content-type":{"type":"string","description":"text/csv; charset=utf-8"}}},"400":{"description":"Bad request, response contains error message.","schema":{"type":"string"}},"401":{"description":"Not authorized"},"404":{"description":"Not found, if path is wrong (unknown context group or feedback type). Response contains error message.","schema":{"type":"string"}},"500":{"description":"Internal server error that might have multiple reasons, for instance no configured global database. Response contains error message.","schema":{"type":"string"}}},"x-ox-examples":{"Example":{"description":"Export user feedback from 2016-01-01 to 2016-12-31 as CSV","example":"GET http://localhost:8009/userfeedback/v1/export/default/star-rating-v1?start=1451606400&end=1483228799"}}}},"/userfeedback/v1/export/{context-group}/{type}/raw":{"get":{"security":[{"basicAuth":[]}],"operationId":"exportRAW","tags":["Userfeedback"],"summary":"Exports user feedback","produces":["application/json"],"parameters":[{"in":"path","name":"context-group","description":"The context group identifying the global DB where the feedback is stored.","required":true,"type":"string","default":"default"},{"in":"path","name":"type","description":"The feedback type to send.","required":true,"type":"string","default":"star-rating-v1"},{"name":"start","in":"query","description":"Start time in milliseconds since 1970-01-01 00:00:00 UTC. Only feedback given after this time is sent. If not set, all feedback up to -e is sent.","required":false,"type":"integer","format":"int64"},{"name":"end","in":"query","description":"End time in milliseconds since 1970-01-01 00:00:00 UTC. Only feedback given before this time is sent. If not set, all feedback since -s is sent.","required":false,"type":"integer","format":"int64"}],"responses":{"200":{"description":"A JSON array containing the stored user feedbacks.","schema":{"type":"string"}},"400":{"description":"Bad request, response contains error message.","schema":{"type":"string"}},"401":{"description":"Not authorized"},"404":{"description":"Not found, if path is wrong (unknown context group or feedback type). Response contains error message.","schema":{"type":"string"}},"500":{"description":"Internal server error that might have multiple reasons, for instance no configured global database. Response contains error message.","schema":{"type":"string"}}},"x-ox-examples":{"Example":{"description":"Export user feedback from 2016-01-01 to 2016-12-31","example":"GET http://localhost:8009/userfeedback/v1/export/default/star-rating-v1/raw?start=1451606400&end=1483228799"}}}},"/userfeedback/v1/send/{context-group}/{type}":{"post":{"operationId":"send","tags":["Userfeedback"],"summary":"Exports user feedback and sends exported data to provided recipients","parameters":[{"name":"context-group","in":"path","description":"The context group identifying the global DB where the feedback is stored.","required":true,"type":"string","default":"default"},{"name":"type","in":"path","description":"The feedback type to send.","required":true,"type":"string","default":"star-rating-v1"},{"name":"start","in":"query","description":"Start time in milliseconds since 1970-01-01 00:00:00 UTC. Only feedback given after this time is sent. If not set, all feedback up to -e is sent.","required":false,"type":"integer","format":"int64"},{"name":"end","in":"query","description":"End time in milliseconds since 1970-01-01 00:00:00 UTC. Only feedback given before this time is sent. If not set, all feedback since -s is sent.","required":false,"type":"integer","format":"int64"},{"name":"requestBody","in":"body","description":"JSONObject with fields \"subject\" for custom mail subject, \"body\" for custom mail body, boolean \"compress\" to compress the mail attachment and \"recipients\" (an JSON Array containing address, display name and PGP key (if available) for every recipient).","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Info message about sent status","schema":{"type":"string"}},"400":{"description":"Bad request, response contains error message.","schema":{"type":"string"}},"401":{"description":"Not authorized"},"404":{"description":"Not found, if path is wrong (unknown context group or feedback type). Response contains error message.","schema":{"type":"string"}},"500":{"description":"Internal server error that might have multiple reasons, for instance no configured global database. Response contains error message.","schema":{"type":"string"}}},"x-ox-examples":{"Example":{"description":"Export user feedback from 2016-01-01 to 2016-12-31 and send it via mail to user1@example.org and user2@example.org","example":"POST http://localhost:8009/userfeedback/v1/send/default/star-rating-v1?start=1451606400&end=1483228799 {\"subject\":\"User Feedback\",\n \"body\":\"User Feedback\",\n \"compress\":true,\n \"recipients\": [\n   {\"address\":\"user1@example.org\",\n    \"displayName\":\"User 1\",\n    \"pgp_key\":\"-----BEGIN PGP PUBLIC KEY BLOCK-----\\r\\nVersion: BCPG v1.56....\"},\n   {\"address\":\"user2@example.org\",\n    \"displayName\":\"User 2\",\n    \"pgp_key\":\"-----BEGIN PGP PUBLIC KEY BLOCK-----\\r\\nVersion: BCPG v1.56....\"}\n ]\n}\n"}}}},"/advertisement/v1/config/name":{"put":{"operationId":"putAdvertisementByName","tags":["Advertisement"],"summary":"Sets an advertisement configuration for a given user by name","parameters":[{"name":"name","in":"query","description":"The user's login name","required":true,"type":"string"},{"name":"contextId","in":"query","description":"The context id","required":true,"type":"integer","format":"int64"},{"name":"body","in":"body","description":"JSONObject describing the advertisement configuration","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"The configuration result"},"400":{"description":"Bad request, response contains error message."},"401":{"description":"Not authorized"}}},"delete":{"operationId":"deleteAdvertisementByName","tags":["Advertisement"],"summary":"Remove the current configuration for the user","parameters":[{"name":"name","in":"query","description":"The user's login name","required":true,"type":"string"},{"name":"contextId","in":"query","description":"The context id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"The configuration result"},"400":{"description":"Bad request, response contains error message."},"401":{"description":"Not authorized"}}}},"/advertisement/v1/config/package":{"put":{"operationId":"putAdvertisementByResellerAndPackage","tags":["Advertisement"],"summary":"Sets an advertisement configuration for a package of a reseller","parameters":[{"name":"reseller","in":"query","description":"The reseller name","required":true,"type":"string"},{"name":"package","in":"query","description":"The package name","required":true,"type":"string"},{"name":"body","in":"body","description":"JSONObject describing the advertisement configuration","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"The configuration result"},"400":{"description":"Bad request, response contains error message."},"401":{"description":"Not authorized"}}},"delete":{"operationId":"deleteAdvertisementByResellerAndPackage","tags":["Advertisement"],"summary":"Remove the current configuration for a package of a reseller","parameters":[{"name":"reseller","in":"query","description":"The reseller name","required":true,"type":"string"},{"name":"package","in":"query","description":"The package name","required":true,"type":"string"}],"responses":{"200":{"description":"The configuration result"},"400":{"description":"Bad request, response contains error message."},"401":{"description":"Not authorized"}}}},"/advertisement/v1/config/reseller":{"put":{"operationId":"putAdvertisementByReseller","tags":["Advertisement"],"summary":"Sets all advertisement configurations for a given reseller","parameters":[{"name":"reseller","in":"query","description":"The reseller's name","required":true,"type":"string"},{"name":"body","in":"body","description":"JSONObject describing the advertisement configuration","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"The configuration result"},"400":{"description":"Bad request, response contains error message."},"401":{"description":"Not authorized"}}}},"/advertisement/v1/config/user":{"put":{"operationId":"putAdvertisementByUserId","tags":["Advertisement"],"summary":"Sets an advertisement configuration for a given user by id","parameters":[{"name":"contextId","in":"query","description":"The context id","required":true,"type":"integer","format":"int64"},{"name":"userId","in":"query","description":"The user id","required":true,"type":"integer","format":"int64"},{"name":"body","in":"body","description":"JSONObject describing the advertisement configuration","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"The configuration result"},"400":{"description":"Bad request, response contains error message."},"401":{"description":"Not authorized"}}},"delete":{"operationId":"deleteAdvertisementByUserId","tags":["Advertisement"],"summary":"Remove the current configuration for the user identified by userId","parameters":[{"name":"contextId","in":"query","description":"The context id","required":true,"type":"integer","format":"int64"},{"name":"userId","in":"query","description":"The user id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"The configuration result"},"400":{"description":"Bad request, response contains error message."},"401":{"description":"Not authorized"}}}}}}